Fix and clear code

+++ MK4duo/src/core/commands/gcode/calibrate/m48.h
@@ -118,7 +118,7 @@
 
     printer.setup_for_endstop_or_probe_move();
 
-    double mean = 0.0, sigma = 0.0, min = 99999.9, max = -99999.9, sample_set[n_samples];
+    float mean = 0.0, sigma = 0.0, min = 99999.9, max = -99999.9, sample_set[n_samples];
 
     // Move to the first point, deploy, and probe
     const float t = probe.check_pt(X_probe_location, Y_probe_location, raise_after, verbose_level);
